Modify useMarketsURLControl
