+++ src/pages/LendPage/components/LendPageContent/LendPageContent.tsx
@@ -3,7 +3,7 @@ import { isEmpty } from 'lodash'
 
 import { Loader } from '@banx/components/Loader'
 
-import { useMarketURLControl } from '@banx/store'
+import { useMarketsURLControl } from '@banx/store'
 
 import FilterSection from '../FilterSection'
 import { EmptyList, MarketsList } from './components'
@@ -12,7 +12,7 @@ import { useLendPageContent } from './hooks'
 import styles from './LendPageContent.module.less'
 
 const LendPageContent = () => {
-  const { visibleMarkets, toggleMarketVisibility } = useMarketURLControl()
+  const { visibleMarkets, toggleMarketVisibility } = useMarketsURLControl()
   const { marketsPreview, isLoading, searchSelectParams, sortParams, showEmptyList } =
     useLendPageContent()
