+++ src/store/markets/useSearchMarketsURLControl.ts
@@ -1,15 +1,12 @@
-import { useEffect } from 'react'
-
-import { useLocation, useNavigate } from 'react-router-dom'
 import { create } from 'zustand'
 
-type SelectedMarketsStore = {
+type SeatchSelectedMarketsState = {
   selectedMarkets: string[]
   setSelectedMarkets: (value: string[]) => void
 }
 
-export const useSelectedMarkets = create<SelectedMarketsStore>((set) => {
-  const initialState: SelectedMarketsStore = {
+export const useSearchSelectedMarkets = create<SeatchSelectedMarketsState>((set) => {
+  const initialState: SeatchSelectedMarketsState = {
     selectedMarkets: [],
     setSelectedMarkets: (value) => {
       set(() => {
@@ -28,23 +25,3 @@ export const useSelectedMarkets = create<SelectedMarketsStore>((set) => {
 
   return initialState
 })
-
-export const useSearchMarketsURLControl = () => {
-  const navigate = useNavigate()
-  const location = useLocation()
-
-  const { selectedMarkets, setSelectedMarkets } = useSelectedMarkets()
-
-  useEffect(() => {
-    const queryParams = new URLSearchParams(location.search)
-    queryParams.delete('collections')
-
-    if (selectedMarkets.length > 0) {
-      queryParams.append('collections', selectedMarkets.join(','))
-    }
-
-    navigate({ search: queryParams.toString() })
-  }, [selectedMarkets, navigate, location.search])
-
-  return { selectedMarkets, setSelectedMarkets }
-}
